added tailwind, created header for authorized user, profile page, Share block
