+++ package.json
@@ -7,11 +7,14 @@
     "@emotion/styled": "^11.11.0",
     "@fontsource/roboto": "^5.0.12",
     "@mui/icons-material": "^5.15.14",
+    "@mui/joy": "^5.0.0-beta.32",
     "@mui/material": "^5.15.10",
     "@mui/styled-engine-sc": "^6.0.0-alpha.17",
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "autoprefixer": "^10.4.19",
+    "postcss-cli": "^11.0.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.22.3",
@@ -19,6 +22,7 @@
     "react-swipeable-views": "^0.14.0",
     "styled-components": "^6.1.8",
     "swiper": "^11.1.0",
+    "tailwindcss": "^3.4.3",
     "typescript": "^5.3.3",
     "web-vitals": "^2.1.4"
   },
