+++ src/components/Auth/Login/login.tsx
@@ -1,0 +1,69 @@
+import React, { useState, useContext } from "react";
+import { useNavigate } from "react-router-dom";
+// import API from "../../API/API";
+
+// // Components
+// import Button from "../Button";
+
+// // Styles
+// import { Wrapper } from "./login.styles";
+
+// // Context
+// import { Context } from "../../context/context";
+
+const Login = () => {
+  const [username, setUsername] = useState("");
+  const [password, setPassword] = useState("");
+  const [error, setError] = useState(false);
+
+//   const [_user, setUser] = useContext(Context);
+  const navigate = useNavigate();
+
+  const handleSubmit = async () => {
+    setError(false);
+    // try {
+    //   const requestToken = await API.getRequestToken();
+    //   const sessionId = await API.authenticate(
+    //     requestToken,
+    //     username,
+    //     password
+    //   );
+    //   console.log(sessionId);
+    //   setUser({ sessionId: sessionId.session_id, username });
+
+    //   navigate("/");
+    // } catch (error) {
+    //   setError(true);
+    // }
+  };
+
+  const handleInput = (e) => {
+    const name = e.currentTarget.name;
+    const value = e.currentTarget.value;
+
+    if (name === "username") setUsername(value);
+    if (name === "password") setPassword(value);
+  };
+
+  return (
+    <div>
+      {error && <div className="error">There was an error!</div>}
+      <label>Username:</label>
+      <input
+        type="text"
+        value={username}
+        name="username"
+        onChange={handleInput}
+      />
+      <input
+        type="password"
+        value={password}
+        name="password"
+        onChange={handleInput}
+      />
+      {/* <Button callback={handleSubmit}>Login</Button> */}
+    </div>
+  );
+};
+
+export default Login;
