+++ src/components/Product/InfoItem/index.tsx
