+++ src/components/Product/InfoItem/infoItem.scss
