+++ src/components/Product/PackageInfo/packageInfo.scss
