+++ src/components/Product/Regional/regional.scss
