+++ src/components/common/FAQPanel/index.tsx
@@ -1,0 +1,49 @@
+import * as React from 'react';
+import AccordionGroup from '@mui/joy/AccordionGroup';
+import Accordion from '@mui/joy/Accordion';
+import AccordionDetails from '@mui/joy/AccordionDetails';
+import AccordionSummary from '@mui/joy/AccordionSummary';
+
+export default function AccordionControlled() {
+  const [index, setIndex] = React.useState<number | null>(0);
+  return (
+    <AccordionGroup sx={{ maxWidth: 400 }}>
+      <Accordion
+        expanded={index === 0}
+        onChange={(event, expanded) => {
+          setIndex(expanded ? 0 : null);
+        }}
+      >
+        <AccordionSummary>First question</AccordionSummary>
+        <AccordionDetails>
+          Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod
+          tempor incididunt ut labore et dolore magna aliqua.
+        </AccordionDetails>
+      </Accordion>
+      <Accordion
+        expanded={index === 1}
+        onChange={(event, expanded) => {
+          setIndex(expanded ? 1 : null);
+        }}
+      >
+        <AccordionSummary>Second question</AccordionSummary>
+        <AccordionDetails>
+          Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod
+          tempor incididunt ut labore et dolore magna aliqua.
+        </AccordionDetails>
+      </Accordion>
+      <Accordion
+        expanded={index === 2}
+        onChange={(event, expanded) => {
+          setIndex(expanded ? 2 : null);
+        }}
+      >
+        <AccordionSummary>Third question</AccordionSummary>
+        <AccordionDetails>
+          Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod
+          tempor incididunt ut labore et dolore magna aliqua.
+        </AccordionDetails>
+      </Accordion>
+    </AccordionGroup>
+  );
+}
