+++ src/components/common/ModalWindow/modalWindow.scss
