+++ src/components/common/Slider/slider.scss
