+++ src/components/common/TabPanel/tabPanel.scss
