+++ src/config/config.tsx
@@ -1,5 +1,8 @@
+const IS_AUTH: boolean = true;
+// const IS_AUTH: boolean = false;
+
 const API_URL: string = "https://www.airalo.com/api/v3/";
 const API_URL_STORE: string = "https://www.airalo.com/api/v2/store/";
 const SEARCH_BASE_URL: string = `${API_URL_STORE}search/?q=`;
 
-export { API_URL, API_URL_STORE, SEARCH_BASE_URL };
+export { IS_AUTH, API_URL, API_URL_STORE, SEARCH_BASE_URL };
