--- src/index.scss
@@ -1,5 +1,0 @@
-body {
-  margin: 0 auto;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
