+++ src/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './index.scss';
+import './index.css';
 import App from './App.tsx';
 import reportWebVitals from './reportWebVitals';
 import "@fontsource/roboto";
