+++ src/pages/Profile/CreditCards/index.tsx
@@ -1,0 +1,13 @@
+import * as React from "react";
+import Box from "@mui/material/Box";
+import apiSettings, { Country, Package } from "../../../API/API.tsx";
+// import PackageInfo from "../PackageInfo/index.tsx";
+import TabPanel from "../../../components/common/TabPanel/index.tsx";
+
+export default function CreditCardsPage() {
+  return (
+    <Box sx={{ width: "100%" }}>
+      <h3>Credit Cards</h3>
+    </Box>
+  );
+}
