+++ src/pages/Profile/Money/index.tsx
@@ -1,0 +1,14 @@
+import * as React from "react";
+import Box from "@mui/material/Box";
+import apiSettings, { Country, Package } from "../../../API/API.tsx";
+// import PackageInfo from "../PackageInfo/index.tsx";
+import TabPanel from "../../../components/common/TabPanel/index.tsx";
+
+export default function MoneyPage() {
+
+  return (
+    <Box sx={{ width: "100%" }}>
+      <h3>Money</h3>
+    </Box>
+  );
+}
