+++ src/pages/Profile/Orders/index.tsx
@@ -1,0 +1,13 @@
+import * as React from "react";
+import Box from "@mui/material/Box";
+import apiSettings, { Country, Package } from "../../../API/API.tsx";
+// import PackageInfo from "../PackageInfo/index.tsx";
+import TabPanel from "../../../components/common/TabPanel/index.tsx";
+
+export default function OrderPage() {
+  return (
+    <Box sx={{ width: "100%" }}>
+      <h3>Order</h3>
+    </Box>
+  );
+}
