+++ src/pages/Profile/Settings/settings.scss
@@ -1,0 +1,35 @@
+
+.button.button-delete-outline {
+  background-color: transparent;
+  border: 1px solid #f2222e;
+  color: #f2222e;
+  padding: 16px 20px 15px;
+}
+.button.button-delete-outline:hover {
+  background-color: #f2222e;
+  border-color: #f2222e;
+  color: #fff;
+}
+.button.small {
+  border-radius: 4px;
+  font-size: 0.625rem;
+  font-weight: 700;
+  letter-spacing: 1.5px;
+  min-width: 167.5px;
+  padding: 7px 25px 9px;
+}
+/*! CSS Used from: https://www.airalo.com/_nuxt/css/f47e8b9.css */
+
+.delete-box h3 {
+  color: #4a4a4a;
+  font-size: 0.9375rem;
+  font-weight: 500;
+  letter-spacing: 0;
+  line-height: 1.33333;
+}
+.delete-box p {
+  color: #8a8a8a;
+  font-size: 0.8125rem;
+  letter-spacing: 0;
+  line-height: 1.38461;
+}
